move the journal to a submodule
