+++ bd-resilient-kv/src/lib.rs
@@ -20,82 +20,11 @@ mod tests;
 pub mod double_buffered;
 pub mod in_memory;
 pub mod kv_store;
+pub mod kvjournal;
 pub mod memmapped;
 
-use bd_bonjson::Value;
-use std::collections::HashMap;
-
-/// Callback function type for high water mark notifications.
-///
-/// Called when a buffer usage exceeds the high water mark threshold.
-/// Parameters: (`current_position`, `buffer_size`, `high_water_mark_position`)
-pub type HighWaterMarkCallback = fn(usize, usize, usize);
-
-/// Trait for a key-value journaling system whose data can be recovered up to the last successful
-/// write checkpoint.
-pub trait KVJournal {
-  /// Get the current high water mark position.
-  fn high_water_mark(&self) -> usize;
-
-  /// Check if the high water mark has been triggered.
-  fn is_high_water_mark_triggered(&self) -> bool;
-
-  /// Get the current buffer usage as a percentage (0.0 to 1.0).
-  fn buffer_usage_ratio(&self) -> f32;
-
-  /// Get the time when the journal was initialized (nanoseconds since UNIX epoch).
-  fn get_init_time(&self) -> u64;
-
-  /// Generate a new journal entry recording the setting of a key to a value.
-  ///
-  /// Note: Setting to `Value::Null` will mark the entry for DELETION!
-  ///
-  /// # Errors
-  /// Returns an error if the journal entry cannot be written.
-  fn set(&mut self, key: &str, value: &Value) -> anyhow::Result<()>;
-
-  /// Generate a new journal entry recording the deletion of a key.
-  ///
-  /// # Errors
-  /// Returns an error if the journal entry cannot be written.
-  fn delete(&mut self, key: &str) -> anyhow::Result<()>;
-
-  /// Clear all key-value pairs from the journal.
-  ///
-  /// This is more efficient than calling `delete()` on each key individually
-  /// as it can be implemented as a single operation rather than multiple journal entries.
-  ///
-  /// # Errors
-  /// Returns an error if the clearing operation fails.
-  fn clear(&mut self) -> anyhow::Result<()>;
-
-  /// Get the current state of the journal as a `HashMap`.
-  ///
-  /// # Errors
-  /// Returns an error if the buffer cannot be decoded.
-  fn as_hashmap(&mut self) -> anyhow::Result<HashMap<String, Value>>;
-
-  /// Reinitialize this journal using the data from another journal.
-  ///
-  /// This method clears the current contents and populates this journal with all
-  /// key-value pairs from the other journal. The high water mark is not affected.
-  ///
-  /// # Errors
-  /// Returns an error if the other journal cannot be read or if writing to this journal fails.
-  fn reinit_from(&mut self, other: &mut dyn KVJournal) -> anyhow::Result<()>;
-
-  /// Synchronize changes to persistent storage.
-  ///
-  /// For memory-mapped implementations, this forces changes to be written to disk.
-  /// For in-memory implementations, this is a no-op.
-  ///
-  /// # Errors
-  /// Returns an error if the sync operation fails.
-  fn sync(&self) -> anyhow::Result<()> {
-    // Default implementation is a no-op for non-persistent journals
-    Ok(())
-  }
-}
+// Re-export the main trait and types for convenience
+pub use kvjournal::{KVJournal, HighWaterMarkCallback};
 
 // Re-export the in-memory implementation for convenience
 pub use double_buffered::DoubleBufferedKVJournal;
