+++ bd-resilient-kv/src/memmapped.rs
@@ -6,7 +6,7 @@
 // https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
 
 use crate::in_memory::InMemoryKVJournal;
-use crate::{HighWaterMarkCallback, KVJournal};
+use crate::kvjournal::{KVJournal, HighWaterMarkCallback};
 use bd_bonjson::Value;
 use memmap2::{MmapMut, MmapOptions};
 use std::collections::HashMap;
