+++ bd-resilient-kv/src/tests/double_buffered_selection_test.rs
@@ -7,7 +7,7 @@
 
 // Test cases for double-buffered journal selection logic
 
-use crate::{DoubleBufferedKVJournal, InMemoryKVJournal, KVJournal};
+use crate::{DoubleBufferedKVJournal, InMemoryKVJournal, kvjournal::KVJournal};
 use bd_bonjson::Value;
 
 /// Helper to create a journal with specific data and timestamp
