+++ bd-resilient-kv/src/tests/error_handling_test.rs
@@ -7,7 +7,7 @@
 
 // Test cases for error handling and edge cases
 
-use crate::{InMemoryKVJournal, KVJournal};
+use crate::{InMemoryKVJournal, kvjournal::KVJournal};
 use bd_bonjson::Value;
 
 #[test]
