+++ bd-resilient-kv/src/tests/memmapped_test.rs
@@ -5,7 +5,7 @@
 // LICENSE file or at:
 // https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
 
-use crate::{KVJournal, MemMappedKVJournal};
+use crate::{kvjournal::KVJournal, MemMappedKVJournal};
 use bd_bonjson::Value;
 use tempfile::NamedTempFile;
